# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/statistics/summary": {
      "get": {
        "operationId": "StatisticsController_summary",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Aggregated summary (defaults to current month)",
        "tags": [
          "statistics"
        ]
      }
    },
    "/statistics/daily": {
      "get": {
        "operationId": "StatisticsController_daily",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyStatRowDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Per-day series for charts (gap-filled)",
        "tags": [
          "statistics"
        ]
      }
    }
  },
  "info": {
    "title": "Phone & Accessory Shop API",
    "description": "Backend for a small retail shop: phones, accessories, sales, debts, expenses and statistics. Multi-tenant — all data scoped by shop.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "StatRangeDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "example": "2026-08-01"
          },
          "to": {
            "type": "string",
            "example": "2026-08-31"
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
      "PhoneStatsDto": {
        "type": "object",
        "properties": {
          "purchasedCount": {
            "type": "number"
          },
          "purchasedAmount": {
            "type": "number"
          },
          "soldCount": {
            "type": "number"
          },
          "soldAmount": {
            "type": "number"
          },
          "soldCostAmount": {
            "type": "number"
          },
          "profit": {
            "type": "number"
          },
          "returnedCount": {
            "type": "number"
          },
          "returnedAmount": {
            "type": "number"
          },
          "soldOnDebtCount": {
            "type": "number"
          },
          "soldOnDebtAmount": {
            "type": "number"
          },
          "inStockCount": {
            "type": "number",
            "description": "Current snapshot, NOT range-bound"
          },
          "inStockCostAmount": {
            "type": "number",
            "description": "Current snapshot, NOT range-bound"
          }
        },
        "required": [
          "purchasedCount",
          "purchasedAmount",
          "soldCount",
          "soldAmount",
          "soldCostAmount",
          "profit",
          "returnedCount",
          "returnedAmount",
          "soldOnDebtCount",
          "soldOnDebtAmount",
          "inStockCount",
          "inStockCostAmount"
        ]
      },
      "AccessoryStatsDto": {
        "type": "object",
        "properties": {
          "purchasedQty": {
            "type": "number"
          },
          "purchasedAmount": {
            "type": "number"
          },
          "soldQty": {
            "type": "number"
          },
          "soldAmount": {
            "type": "number"
          },
          "soldCostAmount": {
            "type": "number"
          },
          "profit": {
            "type": "number"
          },
          "returnedQty": {
            "type": "number"
          },
          "returnedAmount": {
            "type": "number"
          },
          "remainingQty": {
            "type": "number",
            "description": "Current snapshot"
          },
          "remainingCostAmount": {
            "type": "number",
            "description": "Current snapshot"
          }
        },
        "required": [
          "purchasedQty",
          "purchasedAmount",
          "soldQty",
          "soldAmount",
          "soldCostAmount",
          "profit",
          "returnedQty",
          "returnedAmount",
          "remainingQty",
          "remainingCostAmount"
        ]
      },
      "ExpenseStatsDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "count",
          "total"
        ]
      },
      "DebtStatsDto": {
        "type": "object",
        "properties": {
          "openCount": {
            "type": "number"
          },
          "openAmount": {
            "type": "number"
          },
          "overdueCount": {
            "type": "number"
          },
          "overdueAmount": {
            "type": "number"
          },
          "createdInRangeCount": {
            "type": "number"
          },
          "createdInRangeAmount": {
            "type": "number"
          },
          "collectedInRange": {
            "type": "number"
          }
        },
        "required": [
          "openCount",
          "openAmount",
          "overdueCount",
          "overdueAmount",
          "createdInRangeCount",
          "createdInRangeAmount",
          "collectedInRange"
        ]
      },
      "TotalsStatsDto": {
        "type": "object",
        "properties": {
          "grossProfit": {
            "type": "number",
            "description": "phone profit + accessory profit, net of returns"
          },
          "netProfit": {
            "type": "number",
            "description": "grossProfit - expenses"
          },
          "cashIn": {
            "type": "number",
            "description": "sale paidAmount + debts settled within range"
          },
          "cashOut": {
            "type": "number",
            "description": "expenses + phone/accessory purchases within range"
          }
        },
        "required": [
          "grossProfit",
          "netProfit",
          "cashIn",
          "cashOut"
        ]
      },
      "StatisticsSummaryDto": {
        "type": "object",
        "properties": {
          "range": {
            "$ref": "#/components/schemas/StatRangeDto"
          },
          "phones": {
            "$ref": "#/components/schemas/PhoneStatsDto"
          },
          "accessories": {
            "$ref": "#/components/schemas/AccessoryStatsDto"
          },
          "expenses": {
            "$ref": "#/components/schemas/ExpenseStatsDto"
          },
          "debts": {
            "$ref": "#/components/schemas/DebtStatsDto"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsStatsDto"
          }
        },
        "required": [
          "range",
          "phones",
          "accessories",
          "expenses",
          "debts",
          "totals"
        ]
      },
      "DailyStatRowDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-01"
          },
          "salesAmount": {
            "type": "number"
          },
          "profit": {
            "type": "number"
          },
          "expenses": {
            "type": "number"
          },
          "debtCollected": {
            "type": "number"
          }
        },
        "required": [
          "date",
          "salesAmount",
          "profit",
          "expenses",
          "debtCollected"
        ]
      }
    }
  }
}